Communicating between a GRiSP and the Rasp and incorporating it into the main code

+++ I2C_GRiSP/I2C_GRiSP.ino
@@ -1,6 +1,6 @@
 #include <Wire.h>
 
-byte dataArray[4];
+float dataArray[4];
 
 
 void setup()
